Chỉnh sửa về phần order cũng và thêm phương thức lấy tổng thu nhập của từng cửa hàng

+++ project/src/model_pay/PayOption.java
@@ -9,6 +9,6 @@ public interface PayOption {
 	 * @param bill
 	 * @param moneyReceived
 	 */
-	boolean getPay(Bill bill, int moneyReceived);
+	boolean getPay(Bill bill, double moneyReceived);
 
 }
